Revert "Revert "fix type for 'printing' hid_t to string""
This reverts commit ed2b054d0d05a9789685c17036043ecfa80272b5.

+++ src/wrap.c
@@ -301,12 +301,12 @@ void R_init_rhdf5 (DllInfo * winDll) {
 
 SEXP HID_2_CHARSXP(hid_t hid) {
     char tmp_string[21]; 
-    sprintf(tmp_string, "%lld", hid);
+    sprintf(tmp_string, "%lld", (long long) hid);
     return(mkChar(tmp_string));
 }
 
 SEXP HID_2_STRSXP(hid_t hid) {
     char tmp_string[21]; 
-    sprintf(tmp_string, "%lld", hid);
+    sprintf(tmp_string, "%lld", (long long) hid);
     return(mkString(tmp_string));
 }
